pyproject.toml
[project]
name = "bean-tools"
version = "1.1.0"
description = ""
authors = [
    {name = "aleyoscar",email = "mail@aleyoscar.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer (>=0.16.1,<0.17.0)",
    "prompt-toolkit (>=3.0.51,<4.0.0)",
    "ofxparse (>=0.21,<0.22)",
    "beancount (>=3.1.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "bean_tools", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
bean-tools = "bean_tools.cli:app"
